Update main to P3149R0 as published
* Update some links

Add myself as an author, change Facebook to Meta (Platforms), and fix
the unifex::async_scope link.

* Start work on the introduction

Rewrite section 2 Introduction to reflect my intent for the document.
Start updating section 2.1 Implementation experience to capture what
we've learned about async scopes at Meta, including the evolution from
unifex::v1::async_scope to unifex::v2::async_scope.

* Update introduction and motivation

 * Remove the async RAII section from the intro.
 * Change the first example in section 3.1 to use the API I'm proposing
   instead of the use_resources API
 * Add a briefer version of the same example in the form of a Tony Table
   to highlight the differences.

* Introduction improvements

* Update example 4.1

Move the example in section 4.1 off the _async-resource_ model.

* Tweak section 4.2

Clarify some thoughts on the example of starting work within a UI
framework.

* Update section 4.3

Update the example in section 4.3 to use `unifex::let_value_with` (as
alluded to in the prose), and join the scope before completing. Also
tweak the prose wording since there's no particular reason to expect the
scheduler passed to `foo` to be the `system_scheduler`.

* Update section 4.4

Update the HTTP server example in section 4.4 to use the new API, and
refer to the io_uring-based example in Unifex.

* Flesh out the propose API

This adds a code block the provides the interface for the following:
 - `sender auto nest(sender auto&&, auto&&)`
 - `concept async_scope`
 - `void spawn(sender auto&&, async_scope auto&&, auto = {})`
 - `sender auto spawn_future(sender auto&&, async_scope auto&&, auto = {})`
 - `struct counting_scope`

* Start specifying what execution::nest means

Stil a work in progress.

* Update section 10 Synopsis

I realized that I was writing in section 6 the specification that
belongs in section 10. I've rearranged things so that section 10 now has
an updated API outline (which should maybe be removed from section 6)
and I've spec'd `execution::nest()`, `execution::async_scope<>`, and
`execution::spawn()`.

I need to revisit section 6 to be a prose description of the tools I'm
spec'ing, and finish section 10.

* Update section 5's intro

Change the "Async Scope, usage guide" intro to describe an async scope
as a "bookkeeping policy" and explain what that means.

* Clean up sections 5 and 5.1

Some small editorial changes.

* Rewrite section 5.2

Replace the "specification" of `execution::nest()` that was in section
5.2 with a prose description of the intent.

* Make section 5.3 prose

Change section 5.3 to a prose description of the
`execution::async_scope` concept. Also, move the Lifetime section down
to the end of section 5 where `counting_scope` will be described.

* Tweak wording in section 5, and overhaul 5.4

This change tweaks some of the wording in sections 5.1, 5.2, and 5.3,
and updates section 5.4 to describe `execution::spawn()` in prose.

* More section 5 updates

Tweak the definitions in 5.1 to include a spawn-env and future-env to
make it easier to talk about the environment provided to the nested
senders by the receivers to which they're connected.

Reorganize section 5.4 to separate the environment-related discussion
from the lifecycle of the spawned operation's operation-state.

Write most of section 5.5, with a couple of TODOs left behind. Spent
some time contemplating whether the double-nest that Unifex performs is
necessary and decided against it. The spawned sender *is* nested to
ensure that any work it does can be protected from use-after-free by the
scope, but the future-sender is not nested; I should probably document
these thoughts in the paper to highlight the existence of the choice and
to explain why I chose what I did.

* Update sections 5.4 and 5.5

This change tweaks the wording of how a spawn-receiver's environment
responds to various queries, and finishes the prose description of
spawn_future().

* Start on section 5.6 execution::counting_scope

This change adds some (incomplete) prose description of a counting_scope
and sets up the document structure for describing the type's methods.
The description of `nest()` hasn't changed, yet, so it's not yet in line
with the intended design.

* Update document metadata

I've reserved myself a new paper number at isocpp.org so updating the
metadata to match.

* Fix URL for v2::async_scope

Co-authored-by: Ján Ondrušek <427228+janondrusek@users.noreply.github.com>

* Include Ján's suggested example

Co-authored-by: Ján Ondrušek <ondrusek@fb.com>

* Flesh out the prose description of counting_scope

This change adds an overview of the behaviour of a `counting_scope` and
describes the interplay between its methods, its nest-senders, and its
join-sender.

* Add descriptions of a few counting_scope methods

This change adds prose descriptions for the counting_scope constructor
and destructor, and for nest().

* Update Note to self

* Minor edits to sections 1-5

This change makes minor editorial changes throughout sections 1-5, and
resolves @janondrusek's concerns re: "know" in scare quotes.

* Explain joined(), join_started(), and use_count()

* Complete section 5

Flesh out the prose description of `join()` and add a description of the
nest-sender's responsibility to decrement the scope's count of
outstanding senders.

* Update section 6, delete section 7

Update section 6 to no longer refer to async-functions, and to add some
experience with Unifex's spawn_future() supporting the pipe operator.

Delete section 7 Q&A since I couldn't understand what the intent was.

* Finished?

Finish section 7 on naming; delete sections 8 and 9 (diagrams and
standardese) since I can't make them correspond to section 6 before the
publishing deadline. Add section 8, acknowledgements.

* Tweaks for publishing

Narrow some text so it looks better as a PDF, and update the document
number to a P- document instead of a D- draft.

---------

Co-authored-by: Ján Ondrušek <ondrusek@fb.com>